The Messages are Initiated

+++ src/components/TextFeild.jsx
@@ -4,7 +4,7 @@ const TextFeild = (props) => {
 
     const clicked = (event )=>{
         var data = event.target.value;
-        data = isNaN(parseInt(data))?0:parseInt(data);
+        data = isNaN(parseFloat(data))?0:parseFloat(data);
         props.fn(data);
         console.log("Data " , data);
         console.log("Type of data " , typeof(data))
